lint cleanup

+++ app/src/main/java/me/ccrama/redditslide/Activities/FullscreenImage.java
@@ -74,7 +74,7 @@ public class FullscreenImage extends BaseActivity {
                                         if (bar.getVisibility() == View.GONE) {
                                             bar.setVisibility(View.VISIBLE);
                                         }
-                                        bar.setProgress((int) ((current * 100) / total));
+                                        bar.setProgress((current * 100) / total);
                                         if (current == total) {
 
                                             bar.setVisibility(View.GONE);
@@ -167,7 +167,6 @@ public class FullscreenImage extends BaseActivity {
 
                                         }
 
-                                        ;
                                     });
                         } catch (Exception e) {
                             Log.v("RedditSlide", "COULDN'T DOWNLOAD!");
